Update: rename project.
