+++ app/src/androidTest/java/com/taufikhidayat/ceritakan/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.taufikhidayat.allnewceritaku
+package com.taufikhidayat.ceritakan
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
