--- app/src/main/java/com/taufikhidayat/allnewceritaku/ui/maps/MapsFragment.kt
@@ -1,29 +1,0 @@
-package com.taufikhidayat.allnewceritaku.ui.maps
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.taufikhidayat.allnewceritaku.databinding.FragmentMapsBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MapsFragment : Fragment() {
-
-    private var _binding: FragmentMapsBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentMapsBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
