+++ app/src/main/java/com/taufikhidayat/ceritakan/ui/maps/MapsFragment.kt
@@ -1,24 +1,24 @@
-package com.taufikhidayat.allnewceritaku.ui.story
+package com.taufikhidayat.ceritakan.ui.maps
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.taufikhidayat.allnewceritaku.databinding.FragmentAddStoryBinding
+import com.taufikhidayat.ceritakan.databinding.FragmentMapsBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class AddStoryFragment : Fragment() {
+class MapsFragment : Fragment() {
 
-    private var _binding: FragmentAddStoryBinding? = null
+    private var _binding: FragmentMapsBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentAddStoryBinding.inflate(inflater, container, false)
+        _binding = FragmentMapsBinding.inflate(inflater, container, false)
         return binding.root
     }
