+++ app/src/main/java/com/taufikhidayat/ceritakan/ui/story/AddStoryFragment.kt
@@ -1,29 +1,25 @@
-package com.taufikhidayat.allnewceritaku.ui.welcome
+package com.taufikhidayat.ceritakan.ui.story
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.taufikhidayat.allnewceritaku.databinding.FragmentWelcomeBinding
+import com.taufikhidayat.ceritakan.databinding.FragmentAddStoryBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class WelcomeFragment : Fragment() {
+class AddStoryFragment : Fragment() {
 
-    private var _binding: FragmentWelcomeBinding? = null
+    private var _binding: FragmentAddStoryBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentWelcomeBinding.inflate(inflater, container, false)
+        _binding = FragmentAddStoryBinding.inflate(inflater, container, false)
         return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
     }
 
     override fun onDestroyView() {
